requirements.txt
pyppeteer>=1.0.2
openai>=1.4.0
python-dotenv>=1.0.0
streamlit>=1.30.0
aiohttp>=3.9.1
user-agents>=2.2.0
pytz>=2023.3
schedule>=1.2.1
tqdm>=4.6.0
loguru>=0.7.0
pydantic>=2.0.0
tenacity>=8.2.0
async-timeout>=4.0.0
cryptography>=41.0.0
fake-useragent>=1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0 